refactor: remove async-trait, use LazyLock, i32→i64, no panic on client init
- Remove async-trait dependency: edition 2024 supports native async fn in traits
- LazyLock<Regex> for validate_package_name: compile regex once instead of per call
- version_code, *_sdk_version: i32 → i64 to avoid truncation of large values
- RuStoreDownloader::new() returns Result instead of panicking on client init
- ZIP extraction: stream APK directly to disk instead of buffering in Vec<u8>

+++ src/domain.rs
@@ -8,10 +8,10 @@ pub struct AppInfo {
     pub download_url: String,
     pub package_name: String,
     pub version_name: String,
-    pub version_code: i32,
-    pub min_sdk_version: i32,
-    pub max_sdk_version: i32,
-    pub target_sdk_version: i32,
+    pub version_code: i64,
+    pub min_sdk_version: i64,
+    pub max_sdk_version: i64,
+    pub target_sdk_version: i64,
     pub file_size: u64,
     pub icon_url: String,
 }
@@ -50,7 +50,6 @@ pub enum DomainError {
 }
 
 /// Interface for application repository (to be implemented by infrastructure layer)
-#[async_trait::async_trait]
 pub trait AppRepository {
     async fn get_app_info(&self, package_name: &str) -> Result<AppInfo, DomainError>;
     async fn download_app(&self, app_info: &AppInfo, download_path: &str) -> Result<String, DomainError>;
